+++ src/timeline/TimelineModel.h
@@ -189,6 +189,7 @@ public:
         Q_INVOKABLE void cacheMedia(QString eventId);
         Q_INVOKABLE bool saveMedia(QString eventId) const;
 
+        void updateLastMessage();
         void addEvents(const mtx::responses::Timeline &events);
         template<class T>
         void sendMessage(const T &msg);
@@ -223,6 +224,7 @@ public slots:
                         emit replyChanged(reply_);
                 }
         }
+        void setDecryptDescription(bool decrypt) { decryptDescription = decrypt; }
 
 private slots:
         // Add old events at the top of the timeline.
@@ -257,7 +259,6 @@ private:
                                const std::string &user_id,
                                const mtx::responses::ClaimKeys &res,
                                mtx::http::RequestErr err);
-        void updateLastMessage();
         void readEvent(const std::string &id);
 
         QHash<QString, mtx::events::collections::TimelineEvents> events;
@@ -270,6 +271,7 @@ private:
 
         bool isInitialSync        = true;
         bool paginationInProgress = false;
+        bool decryptDescription   = true;
 
         QString currentId;
         QString reply_;